Rename demo. Rename std helpers. Add README stub
