+++ README.md
@@ -1,0 +1,10 @@
+# Kr8md
+
+Header-only SIMD C++ library
+
+
+# Features
+
+* Supports SSE4.2 and AVX2
+
+
