+++ include/kr8md/std_helpers.h
