+++ src/kr8md_demo.cpp
@@ -5,7 +5,7 @@
 #include <iostream>
 
 #include <kr8md/kr8md.h>
-#include <kr8md/kr8md_std_helpers.h>
+#include <kr8md/std_helpers.h>
 
 using namespace kr8md;
